requirements.txt
# WebSocket and async
websockets>=12.0
aiohttp>=3.9.0
websocket-client>=1.6.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database (optional, for historical data)
sqlalchemy>=2.0.0

# HTTP requests (for API calls)
requests>=2.31.0

# Logging and utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Optional: Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Optional: Twitter API (for alt data strategy)
tweepy>=4.14.0

# Machine learning (for alt data strategy)
scikit-learn>=1.3.0
